Fixing two issues
Fixes the issue encountered with deducting lunar coins and when a network user leaves before a match has started

+++ StatsMod/StatsMod.cs
@@ -67,7 +67,7 @@ namespace StatsMod
 
         private void OnPlayerLeave(NetworkUser networkUser)
         {
-            if (!NetworkServer.active) { return; }
+            if (!NetworkServer.active || networkUser.masterController is null) { return; }
 
             StatsDatabase.RemoveAll((x) => x.BelongsTo(networkUser.masterController));
         }
